feat: :sparkles: 메모룸 페이지 컴포넌트 조합
- HomeHeader, RoomList 컴폰넌트 제거
