--- src/features/home/components/HomeHeader/HomeHeader.stories.tsx
@@ -1,9 +1,0 @@
-import { ComponentMeta, StoryObj } from '@storybook/react';
-
-import HomeHeader from './HomeHeader';
-
-export default {
-  component: HomeHeader,
-} as ComponentMeta<typeof HomeHeader>;
-
-export const Default: StoryObj<typeof HomeHeader> = {};
