--- src/features/home/components/HomeHeader/HomeHeader.types.ts
@@ -1,1 +1,0 @@
-export interface HomeHeaderProps {}
