--- src/features/home/components/HomeHeader/index.ts
@@ -1,1 +1,0 @@
-export { default } from './HomeHeader';
