--- src/features/home/components/index.ts
@@ -1,1 +1,0 @@
-export { default as HomeHeader } from './HomeHeader';
