--- src/features/room/components/RoomList/index.ts
@@ -1,1 +1,0 @@
-export { default } from './RoomList';
